Move MPD backend to avoid conflicts.
